NAX/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
